Fix bug with wrong ROM offset
Fix a bug with how we pass in our sections of ROM. In the iNes file
format, the first 16 bytes is metadata about the catridge. The ROM
memory then starts directly after this header. Our code was correctly
skipping the header with the start index but we failed to add another 16
bytes onto the ending index. Therefore when we started the program we would
panic trying to read the reset vector at 0x3FFC and 0x3FFD. This commit
fixes that ending index.