First pass at switching to Tonic
First pass at switching the gRPC client/server over to Tonic as the old
Rust gRPC trait seems to be unmaintained. This change also implies
pulling in a Tokio as our async runtime. To use Tonic, we needed to move
to the Rust 2021 edition as well.

This change also includes flattening our debugger/grpc crates back into
a single crate. Since starting this project, Cargo does a much better
job handling multiple binaries in a single crate. This change makes
iteration simpler. Perhaps, we can switch back in the future.

Finally, there is still much work to be done with cleanup overall. This
is just the first pass.